some refactoring
external components placed outside the solution scope.
EmguCV and OpenCV has been compiled and is stored statically outside. It
is copied at the build process to the place it needs to be located.
The same for other components except ColorMine.

Plus some improvements on MultipleScatterAndFunctionsRepresentation
class. Some fields has been implemented (read-only) for the ability of
produced image to be modified outside the class.